Change effective RLDS dataset length: only consider primary datasets
Previously, the effective dataset length of an RLDS dataset was
set to equal the number of samples until every dataset in the
mixture finishes one epoch. This change modifies the logic slightly
to only consider the "primary" datasets in the mixture, i.e.,
datasets with sample_weight == 0 (e.g., Bridge).